refactor(arch): Phase 5 — PR #486 리뷰 5차 수정
- B1: AuthSessionCookie/PasswordHasher/PasswordVerifier → shared/lib/auth/types.ts 이동
- B2: BCRYPT_DEFAULT_SALT_ROUNDS @internal 태그 제거
