+++ src/entities/session/lib/types.ts
@@ -12,14 +12,4 @@ export interface ResponseCookie {
     maxAge: number;
 }
 
-/** Interface for hashing plain-text passwords before storage. */
-export interface PasswordHasher {
-    /** Hash a plain-text password and return the storage-safe hash. */
-    hashPassword(password: string): Promise<string>;
-}
-
-/** Interface for comparing a plain-text password against a stored password hash. */
-export interface PasswordVerifier {
-    /** Compare a plain-text password with a stored password hash in constant time. */
-    verifyPassword(password: string, passwordHash: string): Promise<boolean>;
-}
+export type { PasswordHasher, PasswordVerifier } from '@/shared/lib/auth/types';
